Refactor script.js to simplify and break into classes.
