+++ index.html
@@ -331,6 +331,11 @@
                                         <br><small class="text-muted">Code editor with SPARQL syntax highlighting</small>
                                         <br><small class="text-secondary">License: MIT</small>
                                     </li>
+                                    <li class="mb-3">
+                                        <strong>js-yaml</strong> (v4.1.0)
+                                        <br><small class="text-muted">YAML parser and dumper</small>
+                                        <br><small class="text-secondary">License: MIT</small>
+                                    </li>
                                 </ul>
 
                                 <h6 class="mt-4">Backend Components</h6>
@@ -350,9 +355,14 @@
                                         <br><small class="text-muted">Fetch API implementation for Node.js</small>
                                         <br><small class="text-secondary">License: MIT</small>
                                     </li>
-                                    <li>
+                                    <li class="mb-3">
                                         <strong>https-proxy-agent</strong> (v7.0.5)
                                         <br><small class="text-muted">HTTP(s) proxy agent for Node.js</small>
+                                        <br><small class="text-secondary">License: MIT</small>
+                                    </li>
+                                    <li>
+                                        <strong>sparqljs</strong> (v3.7.3)
+                                        <br><small class="text-muted">SPARQL query parser and generator</small>
                                         <br><small class="text-secondary">License: MIT</small>
                                     </li>
                                 </ul>
